[Chore #160619267] Add tests for the controller
